Only install enum34 on python before 3.4
enum34 is a package that backports python 3.4 enum functionality to earlier
versions. However, it only includes the functionality as of 3.4. If other
packages use more recent enum functionality, this will break them. (See #17
for more details.)

The solution is just to install enum34 only when the python version is
pre-3.4.